machine-readable DUL link, download function moved to utils
